refactor: 컴포넌트 및 중복 로직 메서드로 분리
