+++ week3/src/App.jsx
@@ -3,7 +3,7 @@ import GlobalStyle from './style/GlobalStyle';
 import { theme } from './style/theme';
 import { useState } from "react";
 import Header from './Header/Header.jsx';
-import Main from './Main/Main.jsx';
+import GameContainer from './Main/GameContainer.jsx';
 import gameLevel from './Main/LevelData.jsx';
 
 
@@ -13,22 +13,45 @@ function App() {
   const [maxScore, setMaxScore] = useState(gameLevel[0].cardNum);
   const [reset, setReset] = useState(false);
 
-  const clickedReset = () => {
-    setReset(prev => !prev); 
-};
-
   const handleLevelChange = (level) => {
     setCurrentLevel(level);
     const levelData = gameLevel.find(item => item.level === level);
     setMaxScore(levelData.cardNum);
   }
 
+  const handleCurrentLevel = (level) => {
+    setCurrentLevel(level);
+  };
+  
+  const handleCurrentScore = (score) => {
+    setCurrentScore(score);
+  };
+  
+  const handleMaxScore = (score) => {
+    setMaxScore(score);
+  };
+
+  const handleReset = () => {
+    setReset(prev => !prev);
+    setCurrentLevel(gameLevel[0].level);  
+    setCurrentScore(0); 
+    setMaxScore(gameLevel[0].cardNum);  
+  };
+
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyle />
       <main>
-        <Header currentScore={currentScore} maxScore={maxScore} clickedReset={clickedReset} />
-        <Main currentLevel={currentLevel} setCurrentLevel={setCurrentLevel} handleLevelChange={handleLevelChange} currentScore={currentScore} setCurrentScore={setCurrentScore} maxScore={maxScore} setMaxScore={setMaxScore} reset={reset}/>
+        <Header currentScore={currentScore} maxScore={maxScore} handleReset={handleReset} />
+        <GameContainer 
+          currentLevel={currentLevel} 
+          handleCurrentLevel={handleCurrentLevel} 
+          handleLevelChange={handleLevelChange} 
+          currentScore={currentScore} 
+          handleCurrentScore={handleCurrentScore } 
+          maxScore={maxScore} 
+          handleMaxScore={handleMaxScore} 
+          reset={reset}/>
       </main>
       
     </ThemeProvider>
