--- week3/src/Main/MainStyle.jsx
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-export const MainContainer = styled.div`
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-`;
-
